+++ app/src/main/java/com/jin/honey/feature/recipe/data/RecipeEntity.kt
@@ -1,11 +1,11 @@
-package com.jin.honey.feature.food.data.model
+package com.jin.honey.feature.recipe.data
 
+import androidx.room.Entity
 import com.jin.honey.feature.recipe.domain.model.RecipeStep
 
+@Entity(tableName = "recipe", primaryKeys = ["menuName"])
 data class RecipeEntity(
-    val categoryName: String,
-    val menuName: String,
-    val imageUrl: String,
+    val menuName:String,
     val cookingTime: String,
     val recipeStep: List<RecipeStep>,
 )
