+++ app/src/main/java/com/jin/honey/feature/recipe/data/RecipeTrackingDataSource.kt
@@ -1,0 +1,10 @@
+package com.jin.honey.feature.recipe.data
+
+import androidx.room.Dao
+import androidx.room.Insert
+
+@Dao
+interface RecipeTrackingDataSource {
+    @Insert
+    suspend fun syncRecipes(entity: RecipeEntity)
+}
