+++ app/src/main/java/com/jin/honey/feature/recipe/domain/model/RecipePreview.kt
@@ -1,7 +1,6 @@
 package com.jin.honey.feature.recipe.domain.model
 
 import com.jin.honey.feature.food.domain.model.CategoryType
-import com.jin.honey.feature.food.domain.model.Recipe
 
 data class RecipePreview(
     val categoryType: CategoryType,
